Only pass a block to render if one was actually given
This call style/parameter is documented in e.g: https://guides.rubyonrails.org/layouts_and_rendering.html#rendering-an-action-s-template-from-another-controller

In b0cc8bceea5b397816ef4ae61db8ba723ae2f647 the code was changed from _only passing a block to render if one was originally given_ to _always passing a block to render regardless_. This is enough to break the call to render in some situations.

This change only passes a block to render if one was actually given to begin with.